refactor(architecture): rebaseline enforcement, resolve hotspots, and partition dashboard tests
