+++ docs/plans/architecture-enforcement-next-steps.md
@@ -7,7 +7,9 @@ References:
 - `docs/adr/0021-validation-layering-and-parsing-modes.md`
 - `docs/ui-taxonomy-rollout-guideline.md`
 - `docs/reports/dashboard-navbar-alerts-cluster-inspection-2026-03-13.md`
-- `docs/reports/code-report/code_report_20260313_031650.txt`
+- `docs/reports/code-report/code_report_20260313_033126.txt`
+- `docs/reports/architecture-rebaseline-2026-03-13-round2.md`
+- `.github/pull_request_template.md`
 
 ---
 
@@ -53,37 +55,24 @@ Behavior-family expectation for tests:
 
 ## 3) Known Deviations (Current)
 
-These are accepted temporarily and must be revisited with explicit owner tracking.
+This list is mandatory to update when a deviation is created, resolved, reassigned, or reaches review deadline.
 
-1. `src/modules/process/ui/validation/processApi.validation.ts`
-- Category: refactor local
-- Deviation: validation-adjacent file still re-exports API request helpers
-- Risk: mode confusion between UI parsing and transport orchestration
-- Action: split into `validation/*` and `api/*` ownership with explicit naming
+### Active
 
-2. `src/capabilities/dashboard/application/dashboard.navbar-alerts.readmodel.ts`
-- Category: structural hotspot
-- Deviation: concentration of grouping/sorting/message shaping in one unit
-- Risk: maintenance and review friction
-- Action: extract pure strategy helpers, keep composition entrypoint thin
+| Deviation | Category | Owner | Risk | Action | Status | Next review |
+|---|---|---|---|---|---|---|
+| `src/modules/process/ui/screens/DashboardScreen.tsx` | structural-hotspot | Process UI maintainers | UI complexity drift | Continue extraction into View/Hook/Usecase/Mapper blocks | active | Next architecture sprint |
+| `src/capabilities/search/ui/SearchOverlay.panel.tsx` | structural-hotspot | Search capability UI maintainers | complexity and ownership blur | Split semantic regions into focused child components | active | Next architecture sprint |
+| `src/capabilities/dashboard/application/tests/dashboard.operational-summary.readmodel.integration.test.ts` | guideline-checklist | Dashboard capability maintainers | mixed behavior families in one suite | partition by behavior family with extracted helpers | active | Next architecture sprint |
+| `src/modules/tracking/infrastructure/persistence/supabaseTrackingAlertRepository.ts` | structural-hotspot | Tracking infra maintainers | query/mapping concentration | continue decomposition while preserving boundary rules | active | Next architecture sprint |
 
-3. `src/capabilities/dashboard/interface/http/tests/dashboard.controllers.test.ts`
-- Category: guideline/checklist
-- Deviation: high test density with mixed behavior families
-- Risk: slower review and brittle failures
-- Action: partition test file by behavior family
+### Resolved in this round (2026-03-13)
 
-4. `src/modules/process/ui/screens/DashboardScreen.tsx`
-- Category: structural hotspot
-- Deviation: large orchestration/render concentration
-- Risk: UI complexity drift
-- Action: continue extraction into View/Hook/Usecase/Mapper blocks
-
-5. `src/capabilities/search/ui/SearchOverlay.panel.tsx`
-- Category: structural hotspot
-- Deviation: large panel with mixed rendering and interaction concerns
-- Risk: complexity and ownership blur
-- Action: split semantic regions into focused child components
+| Deviation | Resolution | Evidence |
+|---|---|---|
+| `src/modules/process/ui/validation/processApi.validation.ts` re-exported API helpers | split ownership: validation file now only maps UI form -> request input; request helpers consumed from `ui/api/process.api.ts` | `code-report`: 47 -> 22 lines |
+| `src/capabilities/dashboard/application/dashboard.navbar-alerts.readmodel.ts` monolithic shaping/grouping/sorting | extracted pure helpers (`*.grouping.ts`, `*.sorting.ts`, `*.message-contract.ts`, `*.readmodel.shared.ts`), keeping read model as composition entrypoint | `code-report`: 566 -> 120 lines |
+| `src/capabilities/dashboard/interface/http/tests/dashboard.controllers.test.ts` monolithic mixed suite | replaced by behavior-family test files + shared harness helper | monolith removed (409 lines) and split suite introduced |
 
 ---
 
@@ -92,4 +81,3 @@ These are accepted temporarily and must be revisited with explicit owner trackin
 - No new ADR proposal without evidence from at least one completed refactor plus rebaseline comparison.
 - Prefer checklist + local refactor over new formalization when docs already cover ownership.
 - Keep this file updated every sprint architecture review.
-
