+++ src/modules/process/ui/screens/shipment/usecases/acknowledgeShipmentAlert.usecase.ts
@@ -1,4 +1,4 @@
-import { acknowledgeTrackingAlertRequest } from '~/modules/process/ui/validation/processApi.validation'
+import { acknowledgeTrackingAlertRequest } from '~/modules/process/ui/api/process.api'
 
 export async function acknowledgeShipmentAlert(alertId: string): Promise<void> {
   await acknowledgeTrackingAlertRequest(alertId)
