+++ src/modules/process/ui/screens/shipment/usecases/submitCreateProcess.usecase.ts
@@ -1,9 +1,7 @@
+import { createProcessRequest } from '~/modules/process/ui/api/process.api'
 import type { CreateProcessDialogFormData } from '~/modules/process/ui/CreateProcessDialog'
 import { toReadableErrorMessage } from '~/modules/process/ui/screens/shipment/lib/shipmentError.presenter'
-import {
-  createProcessRequest,
-  toCreateProcessInput,
-} from '~/modules/process/ui/validation/processApi.validation'
+import { toCreateProcessInput } from '~/modules/process/ui/validation/processApi.validation'
 import {
   type ExistingProcessConflict,
   parseExistingProcessConflictError,
