+++ src/modules/process/ui/screens/shipment/usecases/unacknowledgeShipmentAlert.usecase.ts
@@ -1,4 +1,4 @@
-import { unacknowledgeTrackingAlertRequest } from '~/modules/process/ui/validation/processApi.validation'
+import { unacknowledgeTrackingAlertRequest } from '~/modules/process/ui/api/process.api'
 
 export async function unacknowledgeShipmentAlert(alertId: string): Promise<void> {
   await unacknowledgeTrackingAlertRequest(alertId)
